// generated/prisma/models/Court.ts
export type CourtModel = runtime.Types.Result.DefaultSelection<Prisma.$CourtPayload>
// ...
export type $CourtPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "Court"
  objects: {
    club: Prisma.$ClubPayload<ExtArgs>
    surface: Prisma.$SurfacePayload<ExtArgs>
    Booking: Prisma.$BookingPayload<ExtArgs>[]
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: number
    clubId: number
    indoor: boolean
    surfaceId: number
  }, ExtArgs["result"]["court"]>
  composites: {}
}